requirements.txt
matplotlib==3.5.1
numpy==1.22.2
Pillow==9.0.1
pytorch-lightning==1.5.10
pytorch-pretrained-biggan==0.1.1
click==8.0.0
streamlit==1.7.0
regex==2022.3.15
omegaconf==2.1.1
einops==0.4.1
scipy==1.8.0
